requirements.txt
Django==5.2.7
djangorestframework==3.20.1   # si sigues usando DRF para la API
pandas==2.2.3
scikit-learn==1.5.2
joblib==1.4.2
numpy==2.1.3
jinja2==3.1.4                  # aunque Django usa su propio template engine, Jinja puede quedar si lo usas
gunicorn==21.2.0  